requirements.txt
discord.py
jishaku
